# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticación",
        "description": "Documentación para la API con autenticación mediante Sanctum y Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/currencies": {
            "get": {
                "tags": [
                    "Monedas"
                ],
                "summary": "Listar todas las monedas",
                "operationId": "73e0a13248f9f6aea618716227eab68a",
                "responses": {
                    "200": {
                        "description": "Listado de monedas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Currency"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Monedas"
                ],
                "summary": "Crear una nueva moneda",
                "operationId": "57e1c3f26ae9c82bd5996ac5bb738c5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Currency"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Moneda creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/currencies/{id}": {
            "get": {
                "tags": [
                    "Monedas"
                ],
                "summary": "Mostrar detalles de una moneda",
                "operationId": "23440d4caad0c601f680b655f9c907f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la moneda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la moneda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Moneda no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Monedas"
                ],
                "summary": "Actualizar una moneda",
                "operationId": "8a397517013e49ac5bf2db204caaa130",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Currency"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Moneda actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Monedas"
                ],
                "summary": "Eliminar una moneda",
                "operationId": "c5d45f11f33d23039acf71db15a4cf31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moneda eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Moneda eliminada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Currency": {
                "title": "Moneda",
                "required": [
                    "name",
                    "symbol",
                    "exchange_rate"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dólar estadounidense"
                    },
                    "symbol": {
                        "type": "string",
                        "example": "$"
                    },
                    "exchange_rate": {
                        "type": "number",
                        "format": "float",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T15:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}